layout/style/test/test_query_condition_matches_element.html

<!doctype html>
<title>Test for CSSContainerRule::QueryConditionMatchesElement</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
<style>
@container my-container (min-width: 0px),
           my-container (height > 0px),
           my-container,
           my-container (width > 200px),
           my-container (height > 200px),
           unknown-container,
           unsupported-func(foo: bar)
{
  div {
    outline: 2px solid hotpink;
  }
}

section {
  width: 100px;
  height: 100px;
  background-color: gold;
  margin: 10px;
}

section:has(.inside-container) {
  container-type: size;
  container-name: my-container;
}
</style>

<section>
  <div class="inside-container">IN</div>
</section>
<div class="outside-container">OUT</div>

<script>
  const sheet = document.querySelector("style").sheet;
  const outsideContainerEl = document.querySelector(".outside-container");
  const insideContainerEl = document.querySelector(".inside-container");
  const cssContainerRule = SpecialPowers.wrap(sheet.cssRules[0]);
  // build an array of the conditions text for logging
  const conditions = Array.from(cssContainerRule.conditions).map(({name, query}) =>
    `${name ? name : ""}${query ? (name ? " " : "") + query: ""}`)

  // Container query selection requires up-to-date layout information.
  document.body.getBoundingClientRect();

  const test = (el, conditionIndex, expected) =>
    is(
      cssContainerRule.queryConditionMatchesElement(el, conditionIndex),
      expected,
      `"${conditions[conditionIndex]}" ${expected ? "matches" : "does not match"} for ${el.outerHTML}`
    );


  // "my-container (min-width: 0px)"
  test(insideContainerEl, 0, true);
  // "my-container (height > 0px)"
  test(insideContainerEl, 1, true);
  // "my-container"
  test(insideContainerEl, 2, true);
  // "my-container (width > 200px)": not matching as the container (section) is only 100px wide
  test(insideContainerEl, 3, false);
  // "my-container (height > 200px)": not matching as the container (section) is only 100px tall
  test(insideContainerEl, 4, false);
  // "unknown-container": not matching as there's not container with such name
  test(insideContainerEl, 5, false);
  // "unsupported-func(foo: bar)": not matching as the function isn't supported in Firefox
  test(insideContainerEl, 6, false);

  // conditions are never matching for element which are not in a container
  test(outsideContainerEl, 0, false);
  test(outsideContainerEl, 1, false);
  test(outsideContainerEl, 2, false);
  test(outsideContainerEl, 3, false);
  test(outsideContainerEl, 4, false);
  test(outsideContainerEl, 5, false);
  test(outsideContainerEl, 6, false);

</script>
